pyproject.toml
[tool.poetry]
name = "django-project-template"
version = "0.1.0"
description = ""
authors = ["Islam Ayman"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
django-environ = "^0.10.0"
whitenoise = "^6.5.0"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
django-cors-headers = "^4.2.0"
django-storages = "^1.13.2"
drf-spectacular = "^0.26.4"
psycopg2-binary = "^2.9.6"
django-solo = "^2.1.0"
django-jsonform = "^2.19.0"
pillow = "^10.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
django-debug-toolbar = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

